_Large systems PR featuring a complete rewrite of parry, a rework to the shield broken status, a re-evaluation of shield parameters, cast-wide shield damage adjustments, and a cast-wide increase in landing lag. These changes aim to increase the strength of universal system mechanics in defense and experiment with overall game feel._

# Parry

_The previous parry system has been an improvement in our opinion, but has major issues that lead it to feel unreliable against many attacks. Successful parry hitlag and endlag are inconsistent, varying on attack strength. This causes it to be difficult to react to successful parries on low strength attacks, and difficult to time punishes on higher strength attacks. The new parry system aims to improve this with consistent hitlag and endlag for all parried attacks:_

## Parry input:
- Maximum parry window: 5F -> 7F
- Parry window is now based on when the parry input enters the buffer, rather than when the parry animation starts
  - *For example: 7F parry window. Player inputs parry on F36 of their move that has an FAF of 39. Parry input expires after F42, causing the resulting parry window to be 4F from F39-F42*
  - *In effect: the parry window is a measurement of the player's input timing, rather than the timing of their actionability. This is similar to the system behind Street Fighter 6's perfect parry window*
- Fixed an issue where parry would not be buffered if not held for the entire buffer window
- Fixed an issue where it was possible to enter parry lag without any parry window for 1F after shield damage

## Successful parry:
- Projectile reflect limit: 7 -> 999
- Attacker hitlag: variable -> 18F
- Defender hitlag: variable -> 12F
- Defender endlag: variable (1F-6F) -> 6F
- May now be canceled into offensive out-of-shield options during endlag
  - USpecial
  - USmash
  - Jump
  - Item toss
  - Grab
- May now be canceled into another parry attempt at any time
- Removed the automatic parry window
- No longer sets the opponent's collision kind for shield
  - Zelda can no longer cancel attacks into DSpecial on parry
  - Pichu can no longer jump cancel USpecial on parry
  - Sonic can no longer jump cancel SSpecial on parry
  - Lucario, Ryu, Ken, and Terry can no longer cancel their attacks on parry
  - Palutena no longer gets colors on parry
  - Palutena Blue NSpecial no longer has less endlag on parry
  - Palutena Purple NSpecial no longer has less endlag on parry
  - Shulk can no longer cancel attacks into NSpecial on parry
  - Incineroar can no longer hitfall on parry
  - Piranha Plant SSpecial no longer has less endlag on parry

https://github.com/HDR-Development/HewDraw-Remix/assets/33226440/30b71094-626b-4d48-b170-a8972d2e5659

https://github.com/HDR-Development/HewDraw-Remix/assets/33226440/c6f8192f-7711-428a-9d1f-2aad06ca65f2

# Shield

_Shield health, decay, and recovery values have been adjusted to align remaining shield health with the amount of shield use, rather than shield damage. The values themselves are roughly similar to those seen in Brawl. What this means is that players will be punished more for keeping their shield up, and less for shielding attacks. This should lower the frequency of "accidental" shield breaks._

_However, many members of the cast have had individual moves overcompensated for this change. This means that they can specifically choose to attempt to break your shield._

_Finally, the shield broken status has been rewritten to remove mashing and to make the reward on shield break more dynamic._

## Stats:
- Shield health: 60 -> 72
- Shield global base damage multiplier: 1.4 -> 1.0
- Shield decay rate: 0.255 -> 0.375
  - Time to decay from full to empty shield: 3.92s -> 3.2s
- Shield regeneration rate: 0.07 -> 0.09
  - Time to regen from empty to full shield: 14.29s -> 13.3s

## Dropping Shield:
- May now be canceled into offensive out-of-shield options during endlag (15F)
  - USpecial
  - USmash
  - Parry
  - Jump
  - Item toss
  - Grab

## Shield Break
- Reworked: Shield-broken players no longer fly into the air!
  - Now, they enter a slow vulnerable crumple state for a set amount of time
  - After the crumple state is over, they're forced into either a neutral getup or roll
  - Mashing does not influence recovery time
- Stats:
  - Shield health after break: 37.5 -> 30
  - FAF: variable (300 max) -> 120
  - Invulnerability: air-time plus 32F -> 24F
- Jigglypuff still dies off the top sorry puff mains

https://github.com/HDR-Development/HewDraw-Remix/assets/33226440/a8aaa4ec-22e0-494b-8192-d2d43244906b

# Shield Damage Adjustments
_Shield damage numbers are added on top of the base damage of the move. For example, a move with 8 damage, but a 10 shield damage stat, will do 18 shield damage._

<details>
  <summary>CLICK ME</summary>

## Mario
- [+] NSpecial (hold): 0 --> 10

## Donkey Kong
- [+] SSpecial: 22 -> 40

## Link
- [+] FAir 1: -3 -> 0
- [+] SSpecial: -1.5 -> 0

## Samus
- [+] NSpecial: -2.5/-7 -> 0

## Yoshi
- [+] DSpecial: 10 -> 25

## Kirby
- [+] DSpecial: 0 -> 10

## Pikachu
- [+] Thunder (connects): 2 -> 15

## Luigi
- [+] NSpecial (hold): 0 --> 11

## Ness
- [+] USmash: 0 -> 8
- [+] DSmash: 0 -> 6

## Captain Falcon
- [+] NSpecial: 0 -> 15 (ground) / 17 (ground turn) / 13 (air) / 15 (air turn)

## Jigglypuff
- [+] SSpecial: 17 -> 34

## Peach
- [+] DSmash: 0 -> 2 (x6 multihits) / 7 (launcher)

## Bowser
- [+] DSpecial: 5 -> 28

## Ice Climbers
- [+] NSpecial (hammer): -4 -> 0

## Dr. Mario:
- [+] NSpecial: (hold): 0 -> 12

## Pichu
- [+] FSmash: 0 -> 1 (x5 multihits) / 1 (launcher)
- [+] DSpecial: 2 -> 12

## Marth
- [+] NSpecial (uncharged): 25 -> 45
- [+] NSpecial (charged): 50 -> 70

## Ganondorf
- [+] DSpecial (air): 10 -> 25
- [+] DSmash (hit 2): 0 -> 9

## Mewtwo
- [+] DSmash (early): 8/4 -> 22/18
- [+] DSmash (late): 8/4 -> 17/13

## Roy
- [+] SSpecial (low 4): 0 -> 2 (x5 multihits) / 2 (launcher)

## Mr. Game & Watch
- [+] SSpecial (3): 25 -> 43

## Pit
- [+] SSpecial (ground): 10 -> 25

## Wario
- [-] NAir: 0.7 -> 0
- [+] DSpecial: 0 -> 12/18

## Ike
- [+] NSpecial: 1 -> 7

## Lucas
- [+] NAir: -1 -> 0
- [+] DSmash: -10 -> 0

## King DeDeDe
- [+] FSmash: 7 -> 29
- [+] DSpecial: 0 -> 24

## Olimar
- [+] FSmash: -10/-7/-4.2 -> 0

## Villager
- [+] DSpecial (axe): 9 -> 23
- [+] FSmash: 11 -> 24

## MegaMan
- [+] UTilt: 0 -> 15 (early) / 12 (late)

## Little Mac
- [+] FSmash (down): 4 -> 17

## Dark Pit
- [+] SSpecial (ground): 4 -> 24

## Lucina
- [+] NSpecial (uncharged): 25 -> 45
- [+] NSpecial (charged): 50 -> 70

## Robin
- [+] DSmash: 0 -> 4 (sword) / 4 (electricity)

## Bowser Jr.
- [+] FSmash: 0 -> 2 (x5 multihits) / 5 (launcher)
- [+] Cannonball (bounce): 8 -> 13

## Duck Hunt
- [+] NSpecial (explosion): -5 -> 0

## Ryu
- [-] Heavy Jab: 2 -> 0
- [+] Heavy Jab 2: -5 -> 0
- [+] FTilt: 8/20 -> 14/36
- [+] FSmash: -4 -> 0
- [+] DSmash: -6/-2 -> 0
- [+] DSpecial (level 1): 15 -> 31
- [+] DSpecial (level 2): 22 -> 42
- [+] DThrow (collateral) 50 -> 80

## Cloud
- [+] SSpecial (limit): 0 -> 10 (launcher)

## Corrin
- [+] FSmash: 0 -> 15
- [+] SSpecial (pin): 0 -> 15 (misses ground) / 5 (hits ground)
- [+] NSpecial (bite): 3 -> 15

## Bayonetta
- [+] NSpecial bullets: -0.6 -> 0
- [+] Bullets (generally): -0.3 -> 0

## Daisy
- [+] DSmash: 0 -> 2 (x6 multihits) / 7 (launcher)

## Ridley
- [+] DSpecial (ground): 0 -> 25

## Simon
- [+] NSpecial: 8 -> 17

## Richter
- [+] FSmash 1/3/6 -> 2 (x5 multihits) / 8 (launcher) / 11 (launcher sweetspot)

## Chrom
- [+] SSpecial (low 4): 0 -> 3 (x4 multihits) / 4 (launcher)

## Dark Samus
- [+] NSpecial: -2.5/-7 -> 0

## King K. Rool
- [+] DSmash: 0 -> 18
- [+] Cannonball: -2 -> 0

## Ken
- [-] Heavy Jab / Jab 2: 2 -> 0
- [+] DSmash: -6/-2 -> 0
- [+] Roundhouse kick: -5 -> 0
- [+] Axe Kick: 5 -> 20
- [+] DThrow (collateral) 50 -> 80

## Incineroar
- [+] FTilt: 0 -> 11
- [+] DSmash: 6 -> 23

## Piranha Plant
- [+] SSpecial: 8 -> 19
- [+] DSpecial: 5 -> 20

## Hero
- [+] FSmash Crit: -5/-6 -> 0
- [+] USmash Crit: -4/-5 -> 0
- [+] Hatchet Man: 20 -> 53
- [+] Crash1 (?): -2/-1 -> 0
- [+] Thwack: -1 -> 0
- [+] USpecial: -1 -> 0

## Banjo & Kazooie
- [+] SSpecial (ground): -0.7/-0.8/-0.9/-1 -> 0

## Terry
- [+] Power Geyser: -4 -> 0

## Byleth
- [+] DSmash: 0 -> 15
- [+] DSpecial (uncharged): 8 -> 24
- [+] DSpecial (charged): 25 -> 58

## Min Min
- [+] FSmash (charged double dragon): 0 -> 10

## Steve
- [+] Anvil: 1 -> 7

## Sephiroth
- [+] DSmash: 5 -> 25

## Kazuya
- [+] Dragon Uppercut: 5 -> 21

## Sora
- [+] DSmash: -7 -> 0

## Mii Brawler
- [+] DSpecial1 (landing): -3 -> 0
- [+] DSpecial2 (kick): 5 -> 14
- [+] DSpecial3 (throw): 5 -> 15

## Mii Gunner
- [+] Super Missile: -5 -> 0

</details>

# Landing Lag Adjustments

<details>
  <summary>CLICK ME</summary>

## Mario
 - [-] Nair: 7 -> 8
 - [-] Fair: 11 -> 13
 - [-] Bair: 7 -> 9
 - [-] Uair: 7 -> 9
 - [=] Dair: 12 -> 12
 - [-] Heavy Landing: 4 -> 5

## Donkey Kong
 - [-] Nair: 10 -> 12
 - [-] Fair: 13 -> 15
 - [-] Bair: 9 -> 11
 - [-] Uair: 13 -> 14
 - [-] Dair: 14 -> 16
 - [-] Heavy Landing: 5 -> 6

## Link
 - [-] Nair: 8 -> 10
 - [-] Fair: 8 -> 12
 - [-] Bair: 7 -> 10
 - [-] Uair: 13 -> 15
 - [-] Dair: 18 -> 20
 - [=] Heavy Landing: 5 -> 5

## Samus
 - [-] Nair: 7 -> 10
 - [-] Fair: 10 -> 14
 - [-] Bair: 9 -> 13
 - [-] Uair: 10 -> 12
 - [-] Dair: 12 -> 14
 - [-] Zair: 17 -> 20
 - [-] SSpecial: 4 -> 10
 - [-] Heavy Landing: 4 -> 6

## Yoshi
 - [-] Nair: 7 -> 9
 - [-] Fair: 10 -> 12
 - [-] Bair: 8 -> 10
 - [-] Uair: 9 -> 11
 - [-] Dair: 14 -> 16
 - [-] Heavy Landing: 4 -> 5

## Kirby
 - [-] Nair: 7 -> 9
 - [-] Fair: 8 -> 11
 - [-] Bair: 8 -> 11
 - [-] Uair: 8 -> 10
 - [-] Dair: 12 -> 14
 - [-] Heavy Landing: 4 -> 5

## Fox
 - [-] Nair: 7 -> 9
 - [-] Fair: 15 -> 16
 - [-] Bair: 10 -> 11
 - [-] Uair: 10 -> 12
 - [-] Dair: 10 -> 12
 - [-] NSpecial: 4 -> 6
 - [-] Heavy Landing: 4 -> 5

## Pikachu
 - [=] Nair: 9 -> 9
 - [-] Fair: 11 -> 12
 - [-] Bair: 9 -> 11
 - [-] Uair: 8 -> 9
 - [=] Dair: 15 -> 15
 - [-] Heavy Landing: 4 -> 5

## Luigi
 - [-] Nair: 7 -> 9
 - [-] Fair: 11 -> 12
 - [-] Bair: 8 -> 9
 - [-] Uair: 7 -> 8
 - [-] Dair: 10 -> 12
 - [-] Heavy Landing: 4 -> 5

## Ness
-   [-] Nair: 7 -> 9
-   [-] Fair: 8 -> 10
-   [-] Bair: 10 -> 12
-   [-] Uair: 9 -> 11
-   [-] Dair: 10 -> 12
 - [-] Heavy Landing: 4 -> 5

## Captain Falcon
 - [-] Nair: 7 -> 10
 - [-] Fair: 12 -> 15
 - [-] Bair: 9 -> 12
 - [-] Uair: 8 -> 10
 - [-] Dair: 12 -> 15
 - [-] Heavy Landing: 4 -> 6

## Jigglypuff
 - [-] Nair: 9 -> 11
 - [-] Fair: 9 -> 11
 - [-] Bair: 10 -> 12
 - [-] Uair: 9 -> 10
 - [-] Dair: 12 -> 13
 - [-] Heavy Landing: 4 -> 5

## Peach 
 - [-] Nair: 8 -> 10
 - [-] Fair: 11 -> 13
 - [-] Bair: 8 -> 10
 - [-] Uair: 7 -> 10
 - [-] Dair: 7 -> 10
 - [-] Float Cancel: 4 -> 6
 - [-] Heavy Landing: 4 -> 6

## Bowser 
 - [-] Nair: 10 -> 12
 - [-] Fair: 13 -> 14
 - [-] Bair: 13 -> 14
 - [-] Uair: 12 -> 13
 - [-] Dair: 21 -> 21
 - [-] Heavy Landing: 4 -> 5

## Ice Climbers 
 - [-] Nair: 8 -> 10
 - [-] Fair: 10 -> 12 
 - [-] Bair: 9 -> 11
 - [-] Uair: 12 -> 14
 - [-] Dair: 13 -> 15
 - [-] Heavy Landing: 4 -> 5

## Sheik  
 - [-] Nair: 8 -> 10
 - [-] Fair: 8 -> 10
 - [-] Bair: 9 -> 11
 - [-] Uair: 12 -> 14
 - [-] Dair: 11 -> 13
 - [-] NSpecial: 9 -> 10
 - [-] Heavy Landing: 4 -> 5

## Zelda 
 - [-] Nair: 9 -> 11
 - [-] Fair: 10 -> 13
 - [-] Bair: 11 -> 13
 - [-] Uair: 12 -> 14
 - [-] Dair: 11 -> 14
 - [-] NSpecial: 7 -> 9
 - [-] SSpecial: 4 -> 6
 - [-] Heavy Landing: 4 -> 6

## Dr. Mario
 - [-] Nair: 8 -> 9
 - [-] Fair: 12 -> 14
 - [-] Bair: 8 -> 10
 - [-] Uair: 8 -> 9
 - [-] Dair: 13 -> 14
 - [-] Heavy Landing: 4 -> 6

## Pichu 
 - [-] Nair: 7 -> 8
 - [-] Fair: 10 -> 12
 - [-] Bair: 13 -> 15
 - [-] Uair: 9 -> 10
 - [-] Dair: 14 -> 15
 - [-] Heavy Landing: 3 -> 5

## Falco 
 - [-] Nair: 7 -> 9
 - [-] Fair: 9 -> 13
 - [-] Bair: 10 -> 11
 - [-] Uair: 9 -> 11
 - [-] Dair: 10 -> 13
 - [-] NSpecial: 4 -> 6
 - [=] Heavy Landing: 5 -> 5

## Marth
 - [-] Nair: 8 -> 12
 - [-] Fair: 9 -> 13
 - [-] Bair: 12 -> 14
 - [-] Uair: 8 -> 12
 - [-] Dair: 16 -> 18
 - [-] Heavy Landing: 4 -> 5

## Young Link 
 - [-] Nair: 8 -> 9
 - [-] Fair: 7 -> 10
 - [-] Bair: 9 -> 11
 - [-] Uair: 11 -> 13
 - [-] Dair: 16 -> 18
 - [-] Zair: 12 -> 14
 - [-] Heavy Landing: 4 -> 5

## Ganondorf 
 - [-] Nair: 11 -> 13
 - [-] Fair: 12 -> 14
 - [-] Bair: 12 -> 14
 - [-] Uair: 10 -> 12
 - [-] Dair: 15 -> 17
 - [-] Heavy Landing: 5 -> 6

## Mewtwo
 - [-] Nair: 10 -> 12
 - [-] Fair: 9 -> 11
 - [-] Bair: 12 -> 14
 - [-] Uair: 9 -> 11
 - [-] Dair: 11 -> 13
 - [=] Heavy Landing: 5 -> 5

## Roy 
 - [-] Nair: 10 -> 13
 - [-] Fair: 10 -> 13
 - [-] Bair: 12 -> 16
 - [-] Uair: 8 -> 12
 - [-] Dair: 13 -> 15
 - [-] Heavy Landing: 5 -> 6

## Game & Watch 
 - [-] Nair: 7 -> 10
 - [-] Fair: 12 -> 14
 - [-] Bair: 12 -> 13
 - [=] Uair: 10 -> 10
 - [-] Dair: 10 -> 12
 - [-] NSpecial: 4 -> 6
 - [-] Heavy Landing: 4 -> 5

## Metaknight 
 - [-] Nair: 9 -> 11
 - [-] Fair: 11 -> 13
 - [-] Bair: 13 -> 14
 - [-] Uair: 9 -> 11
 - [-] Dair: 10 -> 12
 - [-] Heavy Landing: 4 -> 5

## Pit 
 - [-] Nair: 9 -> 10
 - [-] Fair: 11 -> 13
 - [-] Bair: 10 -> 13
 - [-] Uair: 11 -> 13
 - [-] Dair: 10 -> 12
 - [=] Heavy Landing: 5 -> 5

## Zero Suit Samus
 - [-] Nair: 7 -> 9 
 - [-] Fair: 9 -> 11
 - [-] Bair: 9 -> 12
 - [-] Uair: 8 -> 10
 - [-] Dair: 9 -> 11
 - [-] Zair: 14 -> 17
 - [-] Heavy Landing: 4 -> 5

## Wario 
 - [-] Nair: 8 -> 10
 - [-] Fair: 7 -> 9
 - [-] Bair: 12 -> 14
 - [-] Uair: 9 -> 11
 - [=] Dair: 29 -> 29
 - [-] Heavy Landing: 4 -> 5

## Snake 
 - [-] Nair: 12 -> 13
 - [-] Fair: 14 -> 16
 - [-] Bair: 10 -> 12
 - [-] Uair: 10 -> 12
 - [-] Dair: 12 -> 14
 - [-] Heavy Landing: 4 -> 6

## Ike 
 - [-] Nair: 9 -> 12
 - [-] Fair: 13 -> 15
 - [-] Bair: 9 -> 13
 - [-] Uair: 12 -> 15
 - [-] Dair: 12 -> 16
 - [-] Heavy Landing: 4 -> 6

## Squirtle 
 - [-] Nair: 8 -> 9
 - [-] Fair: 8 -> 10
 - [-] Bair: 10 -> 12
 - [-] Uair: 7 -> 9
 - [-] Dair: 9 -> 11
 - [-] Heavy Landing: 4 -> 5

## Ivysaur
 - [-] Nair: 10 -> 12
 - [-] Fair: 9 -> 13
 - [-] Bair: 8 -> 12
 - [-] Uair: 10 -> 14
 - [-] Dair: 12 -> 15
 - [-] Heavy Landing: 4 -> 6

## Charizard 
 - [-] Nair: 11 -> 13
 - [-] Fair: 12 -> 14
 - [-] Bair: 12 -> 16
 - [-] Uair: 12 -> 14
 - [-] Dair: 14 -> 16
 - [-] Heavy Landing: 5 -> 6

## Diddy Kong 
 - [-] Nair: 8 -> 10
 - [-] Fair: 11 -> 12
 - [-] Bair: 8 -> 11
 - [-] Uair: 9 -> 11
 - [-] Dair: 12 -> 14
 - [-] Heavy Landing: 4 -> 5

## Lucas 
 - [-] Nair: 9 -> 11
 - [-] Fair: 7 -> 10
 - [-] Bair: 10 -> 13
 - [-] Uair: 7 -> 9
 - [-] Dair: 10 -> 13
 - [-] Zair: 11 -> 13
 - [-] Heavy Landing: 4 -> 5

## Sonic 
 - [-] Nair: 8 -> 10
 - [-] Fair: 11 -> 13
 - [-] Bair: 10 -> 12
 - [-] Uair: 10 -> 12
 - [-] Dair: 16 -> 18
 - [-] Heavy Landing: 4 -> 5

## King Dedede
 - [-] Nair: 11 -> 12
 - [-] Fair: 14 -> 15
 - [-] Bair: 11 -> 12
 - [-] Uair: 12 -> 14
 - [-] Dair: 15 -> 18
 - [+] Heavy Landing: 6 -> 5

## Olimar
 - [-] Nair: 7 -> 9
 - [-] Fair: 7 -> 9
 - [-] Bair: 8 -> 10
 - [-] Uair: 10 -> 12
 - [-] Dair: 12 -> 14
 - [-] NSpecial: variable (min 2) -> 6
 - [-] Heavy Landing: 4 -> 5

## Lucario 
 - [-] Nair: 7 -> 11
 - [-] Fair: 8 -> 12
 - [-] Bair: 12 -> 16
 - [-] Uair: 10 -> 14
 - [-] Dair: 10 -> 14
 - [-] DSpecial: 4 -> 6
 - [-] Heavy Landing: 4 -> 5

## Rob 
 - [-] Nair: 9 -> 11
 - [-] Fair: 7 -> 10
 - [-] Bair: 12 -> 14
 - [-] Uair: 11 -> 13
 - [-] Dair: 13 -> 15
 - [-] Heavy Landing: 4 -> 5

## Toon Link 
 - [-] Nair: 7 -> 9
 - [-] Fair: 9 -> 11
 - [-] Bair: 7 -> 9
 - [-] Uair: 11 -> 13
 - [-] Dair: 17 -> 19
 - [-] Zair: 15 -> 16
 - [-] Heavy Landing: 4 -> 5

## Wolf 
 - [-] Nair: 7 -> 10
 - [-] Fair: 10 -> 12
 - [-] Bair: 10 -> 13
 - [-] Uair: 11 -> 13
 - [-] Dair: 12 -> 15
 - [-] Heavy Landing: 4 -> 5

## Villager 
 - [-] Nair: 7 -> 8
 - [-] Fair: 11 -> 13
 - [-] Bair: 12 -> 14
 - [-] Uair: 11 -> 13
 - [-] Dair: 12 -> 14
 - [-] Heavy Landing: 4 -> 5

## Megaman
 - [-] Nair: 7 -> 8
 - [-] Fair: 10 -> 12
 - [-] Bair: 14 -> 15
 - [=] UAir: 17 -> 17
 - [-] Dair: 13 -> 15
 - [-] Heavy Landing: 4 -> 5

## Wii Fit Trainer
 - [-] Nair: 7 -> 9
 - [-] Fair: 9 -> 11
 - [-] Bair: 10 -> 12
 - [-] Uair: 9 -> 10
 - [-] Dair: 12 -> 14
 - [-] Heavy Landing: 4 -> 5

## Rosalina 
 - [-] Nair: 7 -> 10
 - [-] Fair: 10 -> 12
 - [-] Bair: 10 -> 12
 - [-] Uair: 9 -> 11
 - [-] Dair: 10 -> 12
 - [-] Heavy Landing: 4 -> 6

## Little Mac
 - [-] Nair: 7 -> 9
 - [-] Fair: 7 -> 9
 - [-] Bair: 8 -> 10
 - [-] Uair: 7 -> 9
 - [-] Dair: 10 -> 12
 - [-] Heavy Landing: 4 -> 6

## Greninja 
 - [-] Nair: 9 -> 11
 - [-] Fair: 11 -> 14
 - [-] Bair: 9 -> 11
 - [-] Uair: 12 -> 14
 - [=] DAir: 20 -> 20
 - [-] Heavy Landing: 4 -> 5

## Palutena  
 - [-] Nair: 9 -> 11
 - [-] Fair: 8 -> 10
 - [-] Bair: 12 -> 14
 - [-] Uair: 11 -> 13
 - [-] Dair: 11 -> 14
 - [-] Heavy Landing: 4 -> 5

## Dark Pit
 - [-] Nair: 9 -> 10
 - [-] Fair: 11 -> 13
 - [-] Bair: 10 -> 13
 - [-] Uair: 11 -> 13
 - [-] Dair: 10 -> 12
 - [=] Heavy Landing: 5 -> 5

## Pac-Man 
 - [-] Nair: 7 -> 8
 - [-] Fair: 8 -> 9
 - [-] Bair: 8 -> 10
 - [-] Uair: 8 -> 9
 - [-] Dair: 10 -> 12
 - [-] Heavy Landing: 4 -> 5

## Lucina
 - [-] Nair: 12 -> 15
 - [-] Fair: 9 -> 12
 - [-] Bair: 11 -> 15
 - [-] Uair: 9 -> 13
 - [-] Dair: 12 -> 16
 - [-] Heavy Landing: 4 -> 6

## Robin 
 - [-] Nair: 9 -> 13
 - [-] Fair: 10 -> 14
 - [-] Bair: 11 -> 15
 - [-] Uair: 11 -> 15
 - [-] Dair: 12 -> 15
 - [-] Heavy Landing: 4 -> 6

## Shulk 
 - [-] Nair: 10 -> 13
 - [-] Fair: 12 -> 15
 - [-] Bair: 14 -> 17
 - [-] Uair: 11 -> 14
 - [-] Dair: 13 -> 16
 - [-] Heavy Landing: 4 -> 6

## Bowser Jr. 
 - [-] Nair: 7 -> 10
 - [-] Fair: 9 -> 12
 - [-] Bair: 10 -> 13
 - [-] Uair: 7 -> 9
 - [-] Dair: 12 -> 14
 - [-] Heavy Landing: 4 -> 5

## Duck Hunt 
 - [-] Nair: 7 -> 9
 - [-] Fair: 8 -> 10
 - [-] Bair: 10 -> 12
 - [-] Uair: 9 -> 11
 - [-] Dair: 12 -> 14
 - [-] Heavy Landing: 4 -> 5

## Ryu  
 - [-] Nair: 7 -> 9
 - [-] Fair: 9 -> 12
 - [-] Bair: 10 -> 14
 - [-] Uair: 10 -> 12
 - [-] Dair: 12 -> 15
 - [-] Heavy Landing: 4 -> 5

## Cloud 
 - [-] Nair: 9 -> 13
 - [-] Fair: 13 -> 16
 - [-] Bair: 10 -> 13
 - [-] Uair: 11 -> 14
 - [-] Dair: 15 -> 18
 - [-] Heavy Landing: 4 -> 6

## Corrin 
 - [-] Nair: 10 -> 13
 - [-] Fair: 10 -> 13
 - [-] Bair: 14 -> 16
 - [-] Uair: 12 -> 14
 - [-] Dair: 12 -> 13
 - [-] Heavy Landing: 4 -> 6

## Bayonetta
 - [-] Nair: 8 -> 10
 - [-] Fair: 10 -> 12
 - [-] Bair: 10 -> 13
 - [-] Uair: 8 -> 11
 - [-] Dair: 18 -> 19
 - [-] Heavy Landing: 4 -> 5

## Inkling 
 - [-] Nair: 7 -> 9
 - [-] Fair: 9 -> 11
 - [-] Bair: 8 -> 10
 - [-] Uair: 8 -> 10
 - [-] Dair: 11 -> 13
 - [-] Heavy Landing: 4 -> 5

## Daisy
 - [-] Nair: 8 -> 9
 - [-] Fair: 12 -> 13
 - [-] Bair: 8 -> 10
 - [-] Uair: 8 -> 9
 - [-] Dair: 12 -> 15
 - [-] Heavy Landing: 4 -> 5

## Ridley 
 - [-] Nair: 10 -> 12
 - [-] Fair: 11 -> 12
 - [-] Bair: 13 -> 14
 - [-] Uair: 12 -> 14
 - [-] Dair: 14 -> 16
 - [-] Heavy Landing: 4 -> 5

## Simon 
 - [-] Nair: 8 -> 11
 - [-] Fair: 10 -> 15
 - [-] Bair: 10 -> 15
 - [-] Uair: 10 -> 14
 - [-] Dair: 18 -> 20
 - [-] SSpecial: 11 -> 14
 - [-] Heavy Landing: 4 -> 6

## Richter
 - [-] Nair: 9 -> 12
 - [-] Fair: 10 -> 14
 - [-] Bair: 12 -> 14
 - [-] Uair: 8 -> 10
 - [-] Dair: 18 -> 20
 - [-] Heavy Landing: 4 -> 6

…